Convert dev to LongTensor

+++ codes_zcj/inputters/vanilla.py
@@ -189,7 +189,7 @@ class FeatureDataset(Dataset):
         attention_mask = pad_sequence([torch.tensor([1.] * f.input_length, dtype=torch.float) for f in features],
                                       batch_first=True, padding_value=0.)
         input_length = torch.tensor([f.input_length for f in features], dtype=torch.long)
-        dev = torch.tensor([f.dev for f in features], dtype=torch.int)
+        dev = torch.tensor([f.dev for f in features], dtype=torch.long)
         if not infer:
             decoder_input_ids = pad_sequence([torch.tensor(f.decoder_input_ids, dtype=torch.long) for f in features],
                                              batch_first=True, padding_value=pad)
